# static/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Takas Backend API",
    "description": "İkinci el takas mobil uygulaması için REST API",
    "version": "1.0.0",
    "contact": {
      "name": "Takas Backend Team",
      "email": "info@takas.com"
    }
  },
  "servers": [
    {
      "url": "http://localhost:8000/api/v1",
      "description": "Development server"
    }
  ],
  "components": {
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
...
      "Trade": {
        "type": "object",
        "properties": {
          "id": {"type": "integer"},
          "product_id": {"type": "integer"},
          "user_id": {"type": "integer"},
          "status": {"type": "string"},
          "message": {"type": "string"},
          "created_at": {"type": "string", "format": "date-time"}
        }
      },
...
    }
  },
  "paths": {
...
    "/trades": {
      "get": {
        "tags": ["Trades"],
        "summary": "Takas tekliflerini listele",
        "description": "Kullanıcının takas tekliflerini listeler",
        "security": [{"BearerAuth": []}],
        "parameters": [
          {"name": "page", "in": "query", "schema": {"type": "integer"}, "description": "Sayfa numarası"},
          {"name": "per_page", "in": "query", "schema": {"type": "integer"}, "description": "Sayfa başına teklif sayısı"}
        ],
        "responses": {
          "200": {
            "description": "Takas teklifleri",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "trades": {"type": "array", "items": {"$ref": "#/components/schemas/Trade"}},
                    "total": {"type": "integer"},
                    "pages": {"type": "integer"},
                    "current_page": {"type": "integer"}
                  }
                }
              }
            }
          },
          "401": {"description": "Yetkilendirme gerekli"}
        }
      },
      "post": {
        "tags": ["Trades"],
        "summary": "Yeni takas teklifi oluştur",
        "description": "Bir ürün için takas teklifi oluşturur",
        "security": [{"BearerAuth": []}],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["product_id"],
                "properties": {
                  "product_id": {"type": "integer", "example": 1},
                  "message": {"type": "string", "example": "Bu ürünü takas etmek istiyorum"}
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Takas teklifi başarıyla oluşturuldu",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {"type": "string"},
                    "trade": {"$ref": "#/components/schemas/Trade"}
                  }
                }
              }
            }
          },
          "400": {"description": "Geçersiz veri"},
          "401": {"description": "Yetkilendirme gerekli"}
        }
      }
    },
...
  },
...
}